# Conflicts:
#	app/src/main/java/nl/bezorgdirect/mijnbd/api/ApiDataClasses.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/AssignmentActivity.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/AssignmentFinishedFragment.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/CancelAssignmentFragment.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/DeliveringFragment.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/NewAssignmentFragment.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/delivery/RetrievingFragment.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/helpers/RetrofitHelper.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/history/MyBDHistory.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/recyclerviews/HistoryAdapter.kt
#	app/src/main/java/nl/bezorgdirect/mijnbd/services/NotificationService.kt

+++ app/src/main/java/nl/bezorgdirect/mijnbd/recyclerviews/HistoryAdapter.kt
@@ -72,8 +72,8 @@ class HistoryAdapter (val list: ArrayList<Delivery>, val clicklistener: HistoryL
             holder.distance.text = String.format("%s %s",cont?.resources?.getString(R.string.unknown),cont?.resources?.getString(R.string.lbl_kilometers_short))
         }
         if(item.price != null) {
-            val earnings = round((item.price!! * 100).toDouble()) / 100
-            holder.earnings.text = String.format("%s%2f",cont?.resources?.getString(R.string.lbl_euro),earnings)
+            val earnings = round(item.price!! * 100) / 100
+            holder.earnings.text = String.format("%s%.2f",cont?.resources?.getString(R.string.lbl_euro),earnings)
         }
         else
         {
